refactor: move authorization state to AuthorizationManager
